+++ apollo-router/tests/integration/telemetry/mod.rs
@@ -3,7 +3,6 @@ use std::collections::HashSet;
 
 use opentelemetry::trace::TraceId;
 
-mod client_override;
 #[cfg(any(not(feature = "ci"), all(target_arch = "x86_64", target_os = "linux")))]
 mod datadog;
 mod events;
