+++ apollo-router/tests/integration/telemetry/otlp.rs
@@ -698,6 +698,83 @@ async fn test_priority_sampling_no_parent_propagated() -> Result<(), BoxError> {
     Ok(())
 }
 
+#[tokio::test(flavor = "multi_thread")]
+async fn test_attributes() -> Result<(), BoxError> {
+    if !graph_os_enabled() {
+        return Ok(());
+    }
+    let mock_server = mock_otlp_server(1..).await;
+    let config = include_str!("fixtures/otlp.router.yaml")
+        .replace("<otel-collector-endpoint>", &mock_server.uri());
+    let mut router = IntegrationTest::builder()
+        .telemetry(Telemetry::Otlp {
+            endpoint: Some(format!("{}/v1/traces", mock_server.uri())),
+        })
+        .config(config)
+        .build()
+        .await;
+
+    router.start().await;
+    router.assert_started().await;
+
+    TraceSpec::builder()
+        .services(["client", "router", "subgraph"].into())
+        .attribute("client.name", "foobar")
+        .build()
+        .validate_otlp_trace(
+            &mut router,
+            &mock_server,
+            Query::builder()
+                .traced(true)
+                .header("apollographql-client-name", "foobar")
+                .build(),
+        )
+        .await?;
+    router.graceful_shutdown().await;
+    Ok(())
+}
+
+#[tokio::test(flavor = "multi_thread")]
+async fn test_plugin_overridden_client_name_is_included_in_telemetry() -> Result<(), BoxError> {
+    if !graph_os_enabled() {
+        return Ok(());
+    }
+    let mock_server = mock_otlp_server(1..).await;
+    let config = include_str!("fixtures/otlp_override_client_name.router.yaml")
+        .replace("<otel-collector-endpoint>", &mock_server.uri());
+    let mut router = IntegrationTest::builder()
+        .telemetry(Telemetry::Otlp {
+            endpoint: Some(format!("{}/v1/traces", mock_server.uri())),
+        })
+        .config(config)
+        .build()
+        .await;
+
+    router.start().await;
+    router.assert_started().await;
+
+    // rhai script overrides client.name - no matter what client name we pass via headers, it should
+    // end up equalling the value set in the script (`foo`)
+    for header_value in [None, Some(""), Some("foo"), Some("bar")] {
+        let mut headers = HashMap::default();
+        if let Some(value) = header_value {
+            headers.insert("apollographql-client-name".to_string(), value.to_string());
+        }
+
+        let query = Query::builder().traced(true).headers(headers).build();
+        TraceSpec::builder()
+            .services(["client", "router", "subgraph"].into())
+            .attribute("client.name", "foo")
+            .build()
+            .validate_otlp_trace(&mut router, &mock_server, query)
+            .await
+            .expect(&format!("Failed with header value {header_value:?}"));
+    }
+
+    router.graceful_shutdown().await;
+    Ok(())
+}
+
 struct OtlpTraceSpec<'a> {
     trace_spec: TraceSpec,
     mock_server: &'a MockServer,
